+++ Advent of Code 2023/day3.cpp
@@ -2,7 +2,7 @@
 #include "pch.h"
 using namespace std;
 
-int Solutions::d3_p1(vector<string> input)
+int64_t Solutions::d3_p1(vector<string> input)
 {
 	int summation = 0;
 	int yBound = input.size();
@@ -65,7 +65,7 @@ int Solutions::d3_p1(vector<string> input)
 	return summation;
 }
 
-int Solutions::d3_p2(vector<string> input)
+int64_t Solutions::d3_p2(vector<string> input)
 {
 	long long summation = 0;
 	int yBound = input.size();
