+++ Advent of Code 2023/day7.cpp
@@ -58,7 +58,7 @@ auto rankToStr = [](Rank_Order rank) -> string {
 };
 
 
-int Solutions::d7_p1(vector<string> input)
+int64_t Solutions::d7_p1(vector<string> input)
 {
 	auto cardRank = [](char c) -> int {
 		switch (c) {
@@ -129,7 +129,7 @@ int Solutions::d7_p1(vector<string> input)
 
 // J == Joker, not Jack.
 // don't work, idk
-int Solutions::d7_p2(vector<string> input)
+int64_t Solutions::d7_p2(vector<string> input)
 {
 	auto cardRank = [](char c) -> int {
 		switch (c) {
